requirements.txt
langchain-openai
python-dotenv
openai
flask
twilio
